Separate the crd test from the crud test
